Replace centralized commands on the taxon page vm by commands on the item vm
References #85
The operations on taxon lists (download, select, delete, refresh) were previously directly bound to a command on the TaxonManagementVM instance. That meant, that all "CanExecuteChanged" events caused by a change in one were forwarded to all taxon lists.
Now, the commands are bound to RelayCommand instances for each list which means the change events are more localized.